pyproject.toml
[project]
name = "mlb-data-viewer"
version = "0.1.0"
description = "MLB data extraction pipeline"
requires-python = ">=3.12"
dependencies = [
    "mlb-statsapi",
    "pybaseball>=2.2.0",
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[tool.uv.sources]
mlb-statsapi = { path = "MLB-StatsAPI", editable = true }
